- function separators

+++ source/games/sw/src/game.h
@@ -1982,7 +1982,7 @@ void SOBJ_AlignFloorToPoint(SECTOR_OBJECT* sop, const DVector3& pos);  // morph.
 void ScaleSectorObject(SECTOR_OBJECT* sop); // morph.c
 void MorphTornado(SECTOR_OBJECT* sop);  // morph.c
 void MorphFloor(SECTOR_OBJECT* sop);    // morph.c
-DVector2 ScaleRandomPoint(SECTOR_OBJECT* sop, short k, DAngle ang, const DVector2& pos);    // morph.c
+DVector2 ScaleRandomPoint(SECTOR_OBJECT* sop, int k, DAngle ang, const DVector2& pos);    // morph.c
 
 void CopySectorMatch(int match);  // copysect.c
